requirements.txt
django==4.1.4
gunicorn==20.1.0
djangorestframework==3.14.0
djangorestframework-stubs==3.15.1
markdown==3.4.1
django-filter==22.1
uvicorn==0.20.0
pylint==2.15.8
pylint-django==2.5.3
autopep8==2.0.1
mysqlclient==2.1.1
isort==5.11.2
django-extensions==3.2.1
Werkzeug==2.2.2
requests==2.28.1
transmission-rpc==3.4.0
beautifulsoup4==4.12.3
lxml==5.1.0
redis==5.0.2
